requirements.txt
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2
google-api-python-client>=2.86.0
requests==2.31.0
smtplib
schedule==1.2.0
oauth2client>=4.1.3
gspread==5.10.0
python-dotenv==1.0.0